Implement normalizers instances for google docs and ms word and extract them to separate files. Clean up in contentnormalizer api.
